ros2_k3s_cli/requirements.txt
PyYAML==6.0.2
pydantic==2.11.9
kubernetes==33.1.0
questionary==2.0.1
tabulate==0.9.0